requirements.txt
discord.py
yt-dlp
openai
